requirements.txt
streamlit==1.46.1
plotly==6.2.0
faker==37.4.0
openpyxl==3.1.5
google-generativeai==0.8.5
